+++ src/aux-server/aux-web/aux-projector/interaction/BuilderInteractionManager.ts
@@ -30,6 +30,7 @@ import {
     getContextColor,
     createFile,
     isContext,
+    getFileConfigContexts,
 } from '@casual-simulation/aux-common';
 import { BuilderFileClickOperation } from '../../aux-projector/interaction/ClickOperation/BuilderFileClickOperation';
 import { Physics } from '../../shared/scene/Physics';
@@ -57,6 +58,10 @@ import { GameObject } from '../../shared/scene/GameObject';
 import MiniFile from '../MiniFile/MiniFile';
 import FileTag from '../FileTag/FileTag';
 import FileTable from '../FileTable/FileTable';
+
+import formulaLib, {
+    goToContext,
+} from '../../../../aux-common/Formulas/formula-lib';
 
 export class BuilderInteractionManager extends BaseInteractionManager {
     // This overrides the base class IGameView
@@ -478,6 +483,11 @@ export class BuilderInteractionManager extends BaseInteractionManager {
                     label: minimizedLabel,
                     onClick: () => this._toggleWorkspace(calc, gameObject),
                 });
+
+                actions.push({
+                    label: 'Switch to Player',
+                    onClick: () => this._switchToPlayer(calc, gameObject),
+                });
             }
         }
 
@@ -528,6 +538,24 @@ export class BuilderInteractionManager extends BaseInteractionManager {
                 },
             });
         }
+    }
+
+    private _switchToPlayer(
+        calc: FileCalculationContext,
+        file: ContextGroup3D
+    ) {
+        let contexts = getFileConfigContexts(calc, file.file);
+        let context = contexts[0];
+
+        let url = `${appManager.config.playerBaseUrl}/`;
+
+        // https://auxbuilder.com/
+        //   ^     |     host    |     path           |
+        // simulationId: ''
+        const simulationId = appManager.fileManager.id;
+
+        //window.location.assign(`${url}/${simulationId || 'default'}/${context}`);
+        window.open(`${url}${simulationId || 'default'}/${context}`, '_blank');
     }
 
     private _worldPosToGridPos(